[SimplifyCFG] Fix a cost modeling oversight in branch commoning
The cost modeling was not accounting for the fact we were duplicating the instruction once per predecessor.  With a default threshold of 1, this meant we were actually creating #pred copies.

Adding to the fun, there is *absolutely no* test coverage for this.  Simply bailing for more than one predecessor passes all checked in tests.



git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@341001 91177308-0d34-0410-b5e6-96231b3b80d8