Hashing: use 64-bit seed for hashing on all platforms.
get_execution_seed returns a size_t which varies across platforms, but its
users actually always feed it into a uint64_t role so it makes sense to be
consistent.

Mostly this is just a tidy-up, but it also apparently allows PCH files to be
shared between Clang compilers built for 32-bit and 64-bit hosts.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@341113 91177308-0d34-0410-b5e6-96231b3b80d8