dsymutil: Avoid pruning non-type forward declarations inside DW_TAG_module forward declarations.
Especially with template instantiations, there are legitimate reasons
why for declarations might be emitted into a DW_TAG_module skeleton /
forward-declaration sub-tree, that are not forward declarations in the
sense of that there is a more complete definition over in a .pcm file.

The example in the testcase is a constant DW_TAG_member of a
DW_TAG_class template instatiation.

rdar://problem/43623196

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@341123 91177308-0d34-0410-b5e6-96231b3b80d8