[NFC] Pre-calculate module IR counts in size remarks.
Same as the previous NFC commits in the same vein.

This one introduces a TODO. I'm going to change emitInstrCountChangedRemark
so that it takes in a delta. Since the delta isn't necessary yet, it's not
there. For now, this means that we're calculating the size of the module
twice.

Just done separately to keep the patches small.

4/6

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@341248 91177308-0d34-0410-b5e6-96231b3b80d8