[X86] Add load folding test cases for the addcarryx intrinsic.
We are currently only able to fold a load in operand 1 to ADCX. A load in operand 0 will use the legacy ADC instruction.

Ultimately I want to remove isel support for ADCX, but first I'm going to fix the shortcomings I know of so I can write proper MIR tests to maintain coverage later.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@341744 91177308-0d34-0410-b5e6-96231b3b80d8