NCF: use bit_cast in IntervalMap
This use is sketchy because it basically reinterprets a pointer to another pointer, but right now it's hiding in a union. After this change it'll be easy to grep for bit_cast of a pointer and un-sketch things if we want. This patch therefore obeys the law of conservation of sketch, with minor improvement.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@341900 91177308-0d34-0410-b5e6-96231b3b80d8