[MachineOutliner][NFC] Don't add MBBs with a size < 2 to the search space
The suffix tree won't ever consider sequences with a length less than 2.

Therefore, we really ought to not even consider them in the first place.

Also add a FIXME explaining that this should be defined in terms of the size
in B of an outlined call versus the size in B of the MBB.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@342688 91177308-0d34-0410-b5e6-96231b3b80d8