[llvm-mca] Teach how to track zero registers in class RegisterFile.
This change is in preparation for a future work on improving support for
optimizable register moves.  We already know if a write is from a zero-idiom, so
we can propagate that bit of information to the PRF.  We use an APInt mask to
identify registers that are set to zero.


git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@343307 91177308-0d34-0410-b5e6-96231b3b80d8