[llvm-nm] Fix crash when running with --print-armap on corrupt archives.
error() in llvm-nm intentionally does not return so that the callee can move on to future files/slices. When printing the archive map, this is not currently handled (the caller assumes that error() returns), so processing continues despite there being an error.

Also, change one return to a break, so that symbols can be printed even if the archive map is corrupt.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@344268 91177308-0d34-0410-b5e6-96231b3b80d8