[InlineCost] Remove a dead constant; NFC
My `grep`-fu indicates that this hasn't been used for years. It also no
longer makes much sense to have this flavor of penalty in general, since
a call to a noreturn should mean that we're in a BB that's terminated by
`unreachable`. That case is accounted for by
CallAnalyzer::allowSizeGrowth.


git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@345789 91177308-0d34-0410-b5e6-96231b3b80d8