[X86] Replace an APInt that is guaranteed to be 8-bits with just an 'unsigned'
We're already mixing this APInt with other 'unsigned' variables. This allows us to use regular comparison operators instead of needing to use APInt::ult or APInt::uge. And it removes a later conversion from APInt to unsigned.

I might be adding another combine to this function and this will probably simplify the logic required for that.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@347684 91177308-0d34-0410-b5e6-96231b3b80d8