Adapt gcov to changes in CFE.
The clang frontend no longer emits the current working directory for
DIFiles containing an absolute path in the filename: and will move the
common prefix between current working directory and the file into the
directory: component.

This fixes the GCOV tests in compiler-rt that were broken by the Clang
change.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@348203 91177308-0d34-0410-b5e6-96231b3b80d8